+++ src/record/index.ts
@@ -14,6 +14,7 @@ import {
   recordOptions,
   IncrementalSource,
   listenerHandler,
+  LogRecordOptions,
 } from '../types';
 
 function wrapEvent(e: event): eventWithTime {
@@ -46,6 +47,7 @@ function record<T = eventWithTime>(
     mousemoveWait,
     recordCanvas = false,
     collectFonts = false,
+    recordLog = false,
   } = options;
   // runtime checks for user options
   if (!emit) {
@@ -98,6 +100,37 @@ function record<T = eventWithTime>(
       : _slimDOMOptions
       ? _slimDOMOptions
       : {};
+  const defaultLogOptions: LogRecordOptions = {
+    level: [
+      'assert',
+      'clear',
+      'count',
+      'countReset',
+      'debug',
+      'dir',
+      'dirxml',
+      'error',
+      'group',
+      'groupCollapsed',
+      'groupEnd',
+      'info',
+      'log',
+      'table',
+      'time',
+      'timeEnd',
+      'timeLog',
+      'trace',
+      'warn',
+    ],
+    lengthThreshold: 1000,
+    logger: console,
+  };
+
+  const logOptions: LogRecordOptions = recordLog
+    ? recordLog === true
+      ? defaultLogOptions
+      : Object.assign({}, defaultLogOptions, recordLog)
+    : {};
 
   polyfill();
 
@@ -312,6 +345,16 @@ function record<T = eventWithTime>(
                   },
                 }),
               ),
+            logCb: (p) =>
+              wrappedEmit(
+                wrapEvent({
+                  type: EventType.IncrementalSnapshot,
+                  data: {
+                    source: IncrementalSource.Log,
+                    ...p,
+                  },
+                }),
+              ),
             blockClass,
             blockSelector,
             ignoreClass,
@@ -322,6 +365,7 @@ function record<T = eventWithTime>(
             recordCanvas,
             collectFonts,
             slimDOMOptions,
+            logOptions,
           },
           hooks,
         ),
